feat: camada infra api-search

+++ API-Search/Domain/Abstractions/PagedResult.cs
@@ -3,12 +3,12 @@
 public class PagedResult<T>
 {
     public IEnumerable<T> Items { get; set; } = Enumerable.Empty<T>();
-    public int TotalItems { get; set; }
+    public long TotalItems { get; set; }
     public int PageNumber { get; set; }
     public int PageSize { get; set; }
-    public long TotalPages { get; set; }
+    public int TotalPages { get; set; }
 
-    public PagedResult(IEnumerable<T> items, int totalItems, int pageNumber, int pageSize)
+    public PagedResult(IEnumerable<T> items, long totalItems, int pageNumber, int pageSize)
     {
         Items = items;
         TotalItems = totalItems;
